+++ src/renderer/inline.rs
